+++ glide/confidence_sequences/__init__.py
@@ -1,0 +1,7 @@
+from glide.confidence_sequences.base import ConfidenceSequence
+from glide.confidence_sequences.empirical_bernstein import EmpiricalBernsteinConfidenceSequence
+
+__all__ = [
+    "ConfidenceSequence",
+    "EmpiricalBernsteinConfidenceSequence",
+]